apiserver/requirements/common.txt
# future==0.18.2
amqp==5.2.0  # Celery dependency
asgiref==3.7.2  # Django dependency
billiard==4.2.0  # Celery dependency
celery==5.3.6
click==8.1.7  # Celery dependency
click-didyoumean==0.3.0  # Celery dependency
click-plugins==1.1.1  # Celery dependency
click-repl==0.3.0  # Celery dependency
distlib==0.3.8  # Virtual env dep
Django==5.0.2
django-cors-headers==4.3.1
django_extensions==4.1
django-filter==23.5
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1  # Virtual env dep
gunicorn==21.2.0
kombu==5.3.5  # Celery dependency
packaging==23.2  # Gunicorn depenency
platformdirs==4.2.0  # Virtual env dep
prompt-toolkit==3.0.43  # Celery dependency
PyJWT==2.8.0  # SimpleJWT dependency
python-dateutil==2.8.2  # Celery dependency
pytz==2024.1  # djangorestframework dependency
redis==5.0.1  # django-redis dependency
six==1.16.0  # Celery dependency
sqlparse==0.4.4  # Django dependency
tzdata==2023.4  # Celery dependency
vine==5.1.0  # Celery dependency
virtualenv==20.25.0
wcwidth==0.2.13  # Celery dependency
